// frontend/src/types.ts
export interface Job {
  Title: string;
  Company: string;
  Experience: string;
  Location: string;
  Description: string;
  Salary: string;
  Link: string;
  Site: string;
  Last_Updated: string;
}

export interface Internship {
  Title: string;
  Company: string;
  Experience: string;
  Location: string;
  Description: string;
  Salary: string;
  Link: string;
  Site: string;
  Last_Updated: string;
}